backend/requirements.txt
flask
python-dotenv
flask-restful
flask-restplus
flask_cors
qrcode
pillow
flask_sqlalchemy
flask-jwt-extended
flask-mail
pymysql